code refinement for better readability and extensibility
