+++ nautobot_golden_config/templates/nautobot_golden_config/content_template.html
@@ -83,7 +83,6 @@
                     </tr>
                 {% endif %}
                 {% if config_features.push and golden_config.intended_config %}
-
                     <tr>
                         <td>Configuration to Push</td>
                         <td>
